src/main/java/com/onebit/sample/hexagonal/application/CustomerUseCase.java
package com.onebit.sample.hexagonal.application;

import com.onebit.sample.hexagonal.domain.Customer;
import com.onebit.sample.hexagonal.infra.inputport.CustomerInputPort;
import com.onebit.sample.hexagonal.infra.outputport.EntityRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

import java.util.List;
import java.util.UUID;

@Component
@RequiredArgsConstructor
public class CustomerUseCase implements CustomerInputPort {

    final EntityRepository entityRepository;

    @Override
    public Customer createCustomer(String name, String country) {
        Customer customer = Customer.builder()
            .id( UUID.randomUUID().toString() )
            .name( name )
            .country( country )
            .build();

        return entityRepository.save( customer );
    }

    @Override
    public Customer getById(String customerId) {
        return entityRepository.getById( customerId, Customer.class );
    }

    @Override
    public List<Customer> getAll() {
        return entityRepository.getAll( Customer.class );
    }

}
